Bug 510965 - Indexer getting stuck in infinite loop freeing large chunk
The address space was getting converted to an int and overflowing. Fixed
by casting to (long) whenever a chunk number is converted to an address.

+++ org.eclipse.jdt.core/search/org/eclipse/jdt/internal/core/nd/Nd.java
@@ -662,6 +662,6 @@ public final class Nd {
 	}
 
 	public boolean isValidAddress(long address) {
-		return address > 0 && address < getDB().getChunkCount() * Database.CHUNK_SIZE;
+		return address > 0 && address < (long) getDB().getChunkCount() * Database.CHUNK_SIZE;
 	}
 }
